requirements.txt
lxml==4.6.2
numpy==1.19.4
pandas==1.1.5
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
pyyaml==5.4.1
